Update header version

+++ ThorSerialize/Serialize.h
@@ -47,12 +47,12 @@ namespace ThorsAnvil
 template<typename T>
 T scanValue(char const* buffer, char** end);
 
-template<>  inline short int               scanValue<short int>(char const* buffer, char** end)               {return std::strtol(buffer, end, 10);}
+template<>  inline short int               scanValue<short int>(char const* buffer, char** end)               {return static_cast<short int>(std::strtol(buffer, end, 10));}
 template<>  inline int                     scanValue<int>(char const* buffer, char** end)                     {return std::strtol(buffer, end, 10);}
 template<>  inline long int                scanValue<long int>(char const* buffer, char** end)                {return std::strtol(buffer, end, 10);}
 template<>  inline long long int           scanValue<long long int>(char const* buffer, char** end)           {return std::strtoll(buffer, end, 10);}
 
-template<>  inline unsigned short int      scanValue<unsigned short int>(char const* buffer, char** end)      {return std::strtoul(buffer, end, 10);}
+template<>  inline unsigned short int      scanValue<unsigned short int>(char const* buffer, char** end)      {return static_cast<unsigned short int>(std::strtoul(buffer, end, 10));}
 template<>  inline unsigned int            scanValue<unsigned int>(char const* buffer, char** end)            {return std::strtoul(buffer, end, 10);}
 template<>  inline unsigned long int       scanValue<unsigned long int>(char const* buffer, char** end)       {return std::strtoul(buffer, end, 10);}
 template<>  inline unsigned long long int  scanValue<unsigned long long int>(char const* buffer, char** end)  {return std::strtoull(buffer, end, 10);}
